feat(S11P12A804-106) : 옷 상세 조회 기능 구현

See merge request s11-webmobile1-sub2/S11P12A804!5

+++ backend/clothes/src/main/java/com/ssafy/ozz/clothes/clothes/dto/ClothesCreateRequest.java
@@ -45,8 +45,8 @@ public class ClothesCreateRequest {
     public Clothes toEntity(CategoryLow categoryLow, Long imageFileId, Long userId) {
         return Clothes.builder()
                 .name(name)
-                .size(size.toByte())
-                .fit(fit.toByte())
+                .size(size)
+                .fit(fit)
                 .memo(memo)
                 .brand(brand)
                 .purchaseDate(purchaseDate)
